Go lint
